backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
python-dotenv>=1.0,<2.0
pydantic>=2.10,<3.0
httpx>=0.28,<1.0
pytest>=8.3,<9.0
openai>=2.0,<3.0
azure-search-documents>=11.6,<12.0
azure-identity>=1.25,<2.0
